# perf(div): improve two-limb division on Docker GCC; keep tests stable