标题：perf: 两‑limb 除法（Div/LargeDivisor128）在 Docker/GCC 的性能优化（保持所有用例通过）

动机
- 目标场景：256 位整数除法、两 limb 除数（LargeDivisor128），在 Docker/GCC（CentOS8, GCC 8.5）下为热点；相对 Boost 仍有差距。
- 基线：通用 Knuth D 路径在该场景存在多次 128/64 除法与乘回校正开销，单次除法 ~52.2 ns。

优化思路
- 算法/路径：在两‑limb 专用路径中先用 128 位倒数乘法给出 qhat 初估，再通过少量“向下/向上”修正使 qhat 精确等于 floor(numerator / v1)；随后保持 Knuth D 的第二判断与乘回校正；复用恒等式 qhat*v1 = numerator - rhat，减少额外乘法。
- 位宽专用化：针对 256 位中“两‑limb 除数”的专用实现（函数 div_large_2），不影响其他位宽与路径。
- 可维护性：控制在少量局部变更；保持语义与通用路径等价，新增注释说明修正步骤。

正确性
- 语义不变：严格位宽、补码、移位、比较与模算术行为保持；仅内部估计/修正策略变化。
- 覆盖测试：已覆盖两‑limb 快路径一致性与借位边界的用例；`make test` 全量通过。

基准对比（必填）
- 环境：
  - Docker：CentOS 8 容器镜像 `gint:centos8`；GCC 8.5.0；`-O3 -DNDEBUG`；CMAKE_BUILD_TYPE=Release
  - 本地：Apple clang 17.0.0（macOS），`-O3 -DNDEBUG`；CMAKE_BUILD_TYPE=Release
- 方法：
  - 核心算子：`make bench BENCH_ARGS="--benchmark_min_time=0.2s"`
  - 对比矩阵（256 位）：`make bench-compare-full BENCH_ARGS="--benchmark_min_time=0.2s"`
- 数据（基线 vs 本 PR）：
  - 目标用例（Docker/GCC）Div/LargeDivisor128/gint：~52.2 ns → ~40.8–41.2 ns（~1.27× 加速）
  - 其他 Div/*（Docker+本地）：SmallDivisor32/64、Pow2、SimilarMagnitude/2 未见系统性回退（波动在噪声内）
  - 稳定性：统一固定 `--benchmark_min_time=0.2s`；数据集种子固定；多次复测抖动小。
- 产物（JSON）：
  - Docker：`runs/docker/div_result_compare_opt.json`
  - 本地：`runs/local/div_result_compare_opt.json`
- 提交：
  - 基线：此前 main 的测量（会随时间波动，PR 保留 JSON 作为证据）
  - 本 PR：`$(git rev-parse HEAD)`

总结
- 在不改变语义与通过全部测试的前提下，LargeDivisor128 实测 ~26–28% 提升；其他 Div/* 用例无性能回退。

回归与风险
- 平台差异：使用 `unsigned __int128` 与内建；GCC/Clang/AppleClang 下均已构建与运行；行为与通用路径一致。
- 代码尺寸与可读性：改动集中在两‑limb 专用路径，注释已补充；可维护性可接受。

清单（提交前请逐项勾选）
- [x] 分支名为英文（PR 指南要求）
- [x] 已运行 clang-format（使用仓库根目录 `.clang-format`）
- [x] 新增/更新了 `tests/` 覆盖关键路径，`make test` 通过
- [x] 已提供“变更前/后”基准数据（必填）
- [x] 若更改基准用例，已更新 `bench/` 与 `docs/BENCHMARKS.md`
- [x] 保持 C++11 兼容与 Header-only 特性
- [x] 保持严格位宽与原生类型互操作语义不变